[td] Combine pipeline runs and block runs
