+++ mage_ai/frontend/stories/oracle/elements/Icon.stories.tsx
@@ -54,6 +54,7 @@ import {
   Pause,
   Phone,
   Pipeline,
+  PipelineRun,
   PipelineV2,
   PlayButton,
   PreviewHidden,
@@ -126,6 +127,7 @@ const ICONS = [
   Pause,
   Phone,
   Pipeline,
+  PipelineRun,
   PipelineV2,
   PlayButton,
   PreviewHidden,
